所有页面同步更新收藏状态

+++ library-common/src/main/java/com/pp/common/http/wanandroid/bean/ArticleBean.kt
@@ -78,7 +78,7 @@ data class ArticleBean(
     val userId: Int = 0,
     val visible: Int = 0,
     val zan: Int = 0,
-    val originId: Long? = -1,
+    val originId: Int? = -1,
 ) : Parcelable {
     constructor(parcel: Parcel) : this(
         parcel.readByte() != 0.toByte(),
@@ -116,7 +116,7 @@ data class ArticleBean(
         parcel.readInt(),
         parcel.readInt(),
         parcel.readInt(),
-        parcel.readValue(Long::class.java.classLoader) as? Long
+        parcel.readValue(Int::class.java.classLoader) as? Int
     ) {
     }
 
@@ -171,5 +171,6 @@ data class ArticleBean(
             return arrayOfNulls(size)
         }
     }
+
 
 }
